Remove ga4 tracking javascript
In Whitehall, there are changes so that instead of adding trackers to each individual component/element the tracking will be done by the forms themselves. This is based on the Ga4FinderTracker that's in Publishing Components and expanding it to track changes on non-search forms.

This is generally the approach done by the analytics scripts on GOV.UK: unless there's something specifically different to record, the trackers are added to the parent container rather than the individual trackable things.

Removing individual tracker on the selct with search component